requirements.txt
# Unitree Go2 EDU Development Dependencies

# Core dependencies
numpy>=1.21.0
protobuf>=3.20.0

# Network communication
websocket-client>=1.3.0
requests>=2.28.0

# Data handling
pyyaml>=6.0

# Optional: ROS2 support (uncomment if needed)
# rclpy>=3.3.0
# unitree-go2-ros2>=1.0.0

# Development tools
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

